fix(core): make mini json header compile with libstdc++ (gcc11)
- nlohmann/json.hpp（项目内嵌迷你实现）在类内实例化
  unordered_map<string, json> 成员 variant，libstdc++ 11 不接受
  不完整 mapped type（libc++ 容忍，故 macOS 一直可编）。改为
  shared_ptr<object_t>/<array_t> 间接持有，并手写拷贝构造/赋值
  做深拷贝，保持原有值语义；对外 API 不变。
- dag_serializer: timeout.count() 的 long 赋给 json 在 gcc11 下
  与 operator=(long long)/(size_t) 二义，显式 cast 为 long long。

+++ src/dag_serializer.cpp
@@ -159,7 +159,7 @@ nlohmann::json DAGSerializer::serialize(const DAG& dag) {
         const auto& config = task->config();
         task_json["priority"]    = static_cast<int>(config.priority);
         task_json["max_retries"] = config.max_retries;
-        task_json["timeout_ms"]  = config.timeout.count();
+        task_json["timeout_ms"]  = static_cast<long long>(config.timeout.count());
         task_json["skip_on_fail"] = config.skip_on_fail;
 
         nlohmann::json deps_json = nlohmann::json::array();
